requirements.txt
Django==3.2.7
djangorestframework
markdown
django-filter
Pillow
python-dotenv
mysqlclient
pyyaml
uritemplate
gunicorn
django-heroku
drf-writable-nested